feat: Add form label to AddTextTable component
